changing time bought to double for betatown, implementing variable rate strategy

+++ src/paystation/domain/PayStation.java
@@ -38,7 +38,7 @@ public interface PayStation {
      *
      * @return the number to display on the pay station display
      */
-    int readDisplay();
+    double readDisplay();
 
     /**
      * Buy parking time. Terminate the ongoing transaction and return a parking
